make openai an optional dependency
