+++ src/inspect_ai/agent/_bridge/responses_impl.py
@@ -1,0 +1,581 @@
+import json
+from logging import getLogger
+from time import time
+from typing import Any, cast
+
+from openai.types.responses import (
+    Response,
+    ResponseComputerToolCall,
+    ResponseFunctionToolCall,
+    ResponseInputFileParam,
+    ResponseInputImageParam,
+    ResponseInputItemParam,
+    ResponseInputTextParam,
+    ResponseOutputItem,
+    ResponseOutputMessage,
+    ResponseOutputRefusal,
+    ResponseOutputText,
+    ResponseReasoningItem,
+    ToolParam,
+    WebSearchToolParam,
+)
+from openai.types.responses import (
+    Tool as ResponsesTool,
+)
+from openai.types.responses.response import ToolChoice as ResponsesToolChoice
+from openai.types.responses.response_create_params import (
+    ToolChoice as ResponsesToolChoiceParam,
+)
+from openai.types.responses.response_input_item_param import (
+    Message,
+)
+from pydantic import JsonValue, TypeAdapter
+from shortuuid import uuid
+
+from inspect_ai._util.content import (
+    Content,
+    ContentImage,
+    ContentReasoning,
+    ContentText,
+    ContentToolUse,
+)
+from inspect_ai._util.logger import warn_once
+from inspect_ai.model._call_tools import parse_tool_call
+from inspect_ai.model._chat_message import (
+    ChatMessage,
+    ChatMessageAssistant,
+    ChatMessageSystem,
+    ChatMessageTool,
+    ChatMessageUser,
+)
+from inspect_ai.model._generate_config import GenerateConfig, ResponseSchema
+from inspect_ai.model._openai import (
+    CONTENT_INTERNAL_TAG,
+    _parse_content_with_internal,
+    content_internal_tag,
+)
+from inspect_ai.model._openai_responses import (
+    content_from_response_input_content_param,
+    is_assistant_message_param,
+    is_computer_call_output,
+    is_computer_tool_param,
+    is_function_call_output,
+    is_function_tool_param,
+    is_mcp_tool_param,
+    is_response_computer_tool_call,
+    is_response_function_tool_call,
+    is_response_input_message,
+    is_response_mcp_call,
+    is_response_mcp_list_tools,
+    is_response_output_message,
+    is_response_output_refusal,
+    is_response_output_text,
+    is_response_reasoning_item,
+    is_tool_choice_function_param,
+    is_tool_choice_mcp_param,
+    is_web_search_tool_param,
+    reasoning_from_responses_reasoning,
+    responses_extra_body_fields,
+    responses_model_usage,
+    responses_reasoning_from_reasoning,
+    to_inspect_citation,
+)
+from inspect_ai.model._providers._openai_computer_use import (
+    computer_parmaeters,
+    tool_call_arguments_to_action,
+    tool_call_from_openai_computer_tool_call,
+)
+from inspect_ai.tool._mcp._config import MCPServerConfigHTTP
+from inspect_ai.tool._tool_call import ToolCall
+from inspect_ai.tool._tool_choice import ToolChoice, ToolFunction
+from inspect_ai.tool._tool_info import ToolInfo
+from inspect_ai.tool._tool_params import ToolParams
+from inspect_ai.tool._tools._web_search._web_search import WebSearchProviders
+from inspect_ai.util._json import JSONSchema
+
+from .util import resolve_inspect_model
+
+logger = getLogger(__file__)
+
+
+async def inspect_responses_api_request_impl(
+    json_data: dict[str, Any], web_search: WebSearchProviders
+) -> Response:
+    # resolve model
+    model = resolve_inspect_model(str(json_data["model"]))
+    model_name = model.api.model_name
+
+    # record parallel tool calls
+    parallel_tool_calls = json_data.get("parallel_tool_calls", True)
+
+    # convert openai tools to inspect tools
+    responses_tools: list[ToolParam] = json_data.get("tools", [])
+    tools = [tool_from_responses_tool(tool, web_search) for tool in responses_tools]
+    responses_tool_choice: ResponsesToolChoiceParam | None = json_data.get(
+        "tool_choice", None
+    )
+    tool_choice = tool_choice_from_responses_tool_choice(responses_tool_choice)
+
+    # convert to inspect messages
+    input: list[ResponseInputItemParam] = json_data["input"]
+
+    debug_log("SCAFFOLD INPUT", input)
+
+    messages = messages_from_responses_input(input, tools, model_name)
+    debug_log("INSPECT MESSAGES", messages)
+
+    # run inference
+    output = await model.generate(
+        input=messages,
+        tool_choice=tool_choice,
+        tools=tools,
+        config=generate_config_from_openai_responses(json_data),
+    )
+
+    debug_log("INSPECT OUTPUT", output.message)
+
+    # return response
+    response = Response(
+        id=output.message.id or uuid(),
+        created_at=int(time()),
+        model=model_name,
+        object="response",
+        output=responses_output_items_from_assistant_message(output.message),
+        parallel_tool_calls=parallel_tool_calls,
+        tool_choice=responses_tool_choice_param_to_tool_choice(responses_tool_choice),
+        tools=responses_tool_params_to_tools(responses_tools),
+        usage=responses_model_usage(output.usage),
+    )
+    debug_log("SCAFFOLD RESPONSE", response)
+
+    return response
+
+
+def debug_log(caption: str, o: Any) -> None:
+    # print(caption)
+    # print(to_json_str_safe(o))
+    pass
+
+
+def tool_choice_from_responses_tool_choice(
+    tool_choice: ResponsesToolChoiceParam | None,
+) -> ToolChoice | None:
+    inspect_tool_choice: ToolChoice | None = None
+    if tool_choice is not None:
+        if tool_choice == "auto":
+            inspect_tool_choice = tool_choice
+        elif tool_choice == "none":
+            inspect_tool_choice = tool_choice
+        elif tool_choice == "required":
+            inspect_tool_choice = "any"
+        elif is_tool_choice_function_param(tool_choice):
+            inspect_tool_choice = ToolFunction(name=tool_choice["name"])
+        elif is_tool_choice_mcp_param(tool_choice):
+            if tool_choice["name"] is None:
+                raise RuntimeError(
+                    "MCP server tool choice requires 'name' field for agent bridge"
+                )
+            inspect_tool_choice = ToolFunction(name=tool_choice["name"])
+        elif tool_choice.get("type") == "allowed_tools":
+            raise RuntimeError("ToolChoiceAllowedParam not supported by agent bridge")
+        elif tool_choice.get("type") == "custom":
+            raise RuntimeError("ToolChoiceCustomParam not supported by agent bridge")
+        elif "type" in tool_choice:
+            inspect_tool_choice = ToolFunction(name=str(tool_choice.get("type")))
+
+    return inspect_tool_choice
+
+
+tool_choice_adapter = TypeAdapter[ResponsesToolChoice](ResponsesToolChoice)
+
+
+def responses_tool_choice_param_to_tool_choice(
+    tool_choice: ResponsesToolChoiceParam | None,
+) -> ResponsesToolChoice:
+    if tool_choice is None:
+        return "auto"
+    else:
+        return tool_choice_adapter.validate_python(tool_choice)
+
+
+def tool_from_responses_tool(
+    tool_param: ToolParam, web_search: WebSearchProviders
+) -> ToolInfo:
+    if is_function_tool_param(tool_param):
+        return ToolInfo(
+            name=tool_param["name"],
+            description=tool_param["description"] or tool_param["name"],
+            parameters=ToolParams.model_validate(tool_param["parameters"]),
+        )
+    elif is_web_search_tool_param(tool_param):
+        return ToolInfo(
+            name="web_search",
+            description="web_search",
+            options=responses_web_search_tool_options(tool_param, web_search),
+        )
+    elif is_computer_tool_param(tool_param):
+        return ToolInfo(
+            name="computer",
+            description="computer",
+            # this is a fake parameter def so that we match the check for the
+            # computer tool in maybe_computer_use_preview_tool (openai will
+            # provide its own parmeters internally)
+            parameters=ToolParams(properties={k: k for k in computer_parmaeters()}),  # type: ignore[misc]
+        )
+    elif is_mcp_tool_param(tool_param):
+        allowed_tools = tool_param["allowed_tools"]
+        if isinstance(allowed_tools, dict):
+            raise RuntimeError(
+                "McpAllowedToolsMcpAllowedToolsFilter not supported by agent bridge"
+            )
+        config = MCPServerConfigHTTP(
+            type="sse" if "sse" in tool_param["server_url"] else "http",
+            name=tool_param["server_label"],
+            tools=allowed_tools if isinstance(allowed_tools, list) else "all",
+            url=tool_param["server_url"],
+            headers=tool_param["headers"],
+        )
+        return ToolInfo(
+            name=f"mcp_server_{config.name}",
+            description=f"mcp_server_{config.name}",
+            options=config.model_dump(),
+        )
+    else:
+        raise RuntimeError(f"ToolParam of type {tool_param.get('type')} not supported.")
+
+
+def responses_web_search_tool_options(
+    tool_param: WebSearchToolParam, web_search: WebSearchProviders
+) -> dict[str, Any]:
+    # pass through openai options if there is no special openai config
+    openai_options = web_search.get("openai", False)
+    if (
+        openai_options is True
+        or isinstance(openai_options, dict)
+        and len(openai_options) == 0
+    ):
+        if "user_location" in tool_param or "search_context_size" in tool_param:
+            # this came from the user in the external scaffold. we want
+            # all the fields except the type as our 'web_search' config
+            tool_param = tool_param.copy()
+            del tool_param["type"]  # type: ignore[misc]
+
+            # this came from the inspect agent_bridge() call. we want
+            # to replace it with whatever the user specified in the scaffold.
+            web_search = web_search.copy()
+            web_search["openai"] = tool_param  # type: ignore[typeddict-item]
+
+    return cast(dict[str, Any], web_search)
+
+
+tool_list_adapter = TypeAdapter(list[ResponsesTool])
+
+
+def responses_tool_params_to_tools(tool_params: list[ToolParam]) -> list[ResponsesTool]:
+    return tool_list_adapter.validate_python(tool_params)
+
+
+def generate_config_from_openai_responses(json_data: dict[str, Any]) -> GenerateConfig:
+    # warn for unsupported params
+    def warn_unsupported(param: str) -> None:
+        if param in json_data:
+            warn_once(logger, f"'{param}' option not supported for agent bridge")
+
+    warn_unsupported("background")  # we don't proxy background polling requests
+    warn_unsupported("prompt")  # prompt template
+    warn_unsupported("top_logprobs")  # don't have this yet for responses
+
+    config = GenerateConfig()
+    config.system_message = json_data.get("instructions", None)
+    config.max_tokens = json_data.get("max_output_tokens", None)
+    config.parallel_tool_calls = json_data.get("parallel_tool_calls", None)
+    reasoning = json_data.get("reasoning", None)
+    if reasoning:
+        if "effort" in reasoning:
+            config.reasoning_effort = reasoning["effort"]
+        if "summary" in reasoning:
+            config.reasoning_summary = reasoning["summary"]
+    config.temperature = json_data.get("temperature", None)
+    config.top_p = json_data.get("top_p", None)
+
+    # response format
+    text: dict[str, Any] | None = json_data.get("text", None)
+    if text is not None:
+        format: dict[str, Any] | None = text.get("format", None)
+        if format is not None:
+            if format.get("type", None) == "json_schema":
+                config.response_schema = ResponseSchema(
+                    name=format.get("name", "schema"),
+                    description=format.get("description", None),
+                    json_schema=JSONSchema.model_validate(format.get("schema", {})),
+                    strict=format.get("strict", None),
+                )
+
+    # extra_body params (i.e. passthrough for native responses)
+    extra_body: dict[str, Any] = {}
+    for field in responses_extra_body_fields():
+        if field in json_data:
+            extra_body[field] = json_data[field]
+    if len(extra_body) > 0:
+        config.extra_body = extra_body
+
+    # return config
+    return config
+
+
+def messages_from_responses_input(
+    input: str | list[ResponseInputItemParam],
+    tools: list[ToolInfo],
+    model_name: str | None = None,
+) -> list[ChatMessage]:
+    # TODO: parse out the xml tags, decode it, and set to 'internal'
+    # do this for AssistantInternal and content-internal
+
+    # enture input is a list
+    if isinstance(input, str):
+        input = [
+            Message(
+                type="message",
+                role="user",
+                content=[ResponseInputTextParam(type="input_text", text=input)],
+            )
+        ]
+
+    messages: list[ChatMessage] = []
+    function_calls_by_id: dict[str, str] = {}
+    pending_assistant_message_params: list[ResponseInputItemParam] = []
+
+    def collect_pending_assistant_message() -> None:
+        if len(pending_assistant_message_params) > 0:
+            content: list[Content] = []
+            tool_calls: list[ToolCall] = []
+            for param in pending_assistant_message_params:
+                if is_response_output_message(param):
+                    for output in param["content"]:
+                        text = str(output.get("text", output.get("refusal", "")))
+
+                        asst_content, content_internal = _parse_content_with_internal(
+                            text, CONTENT_INTERNAL_TAG
+                        )
+
+                        if is_response_output_text(output):
+                            content.append(
+                                ContentText(
+                                    text=asst_content,
+                                    internal=content_internal,
+                                    citations=(
+                                        [
+                                            to_inspect_citation(annotation)
+                                            for annotation in output["annotations"]
+                                        ]
+                                        if output.get("annotations", None)
+                                        else None
+                                    ),
+                                )
+                            )
+                        elif is_response_output_refusal(output):
+                            content.append(
+                                ContentText(
+                                    text=output["refusal"],
+                                    refusal=True,
+                                    internal=content_internal,
+                                )
+                            )
+
+                elif is_response_function_tool_call(param):
+                    function_calls_by_id[param["call_id"]] = param["name"]
+                    tool_calls.append(
+                        parse_tool_call(
+                            id=param["call_id"],
+                            function=param["name"],
+                            arguments=param["arguments"],
+                            tools=tools,
+                        )
+                    )
+                elif is_response_computer_tool_call(param):
+                    computer_call = ResponseComputerToolCall.model_validate(param)
+                    tool_calls.append(
+                        tool_call_from_openai_computer_tool_call(computer_call)
+                    )
+
+                elif is_response_reasoning_item(param):
+                    content.append(reasoning_from_responses_reasoning(param))
+                elif is_response_mcp_list_tools(param):
+                    content.append(
+                        ContentToolUse(
+                            tool_type="mcp_list_tools",
+                            id=param["id"],
+                            name="mcp_list_tools",
+                            context=param["server_label"],
+                            arguments="",
+                            result=cast(list[JsonValue], param["tools"]),
+                            error=param.get("error", None),
+                        )
+                    )
+                elif is_response_mcp_call(param):
+                    content.append(
+                        ContentToolUse(
+                            tool_type="mcp_call",
+                            id=param["id"],
+                            name=param["name"],
+                            context=param["server_label"],
+                            arguments=param["arguments"],
+                            result=param.get("output", None),
+                            error=param.get("error", None),
+                        )
+                    )
+                else:
+                    raise RuntimeError(
+                        f"Unexpected assitant message type: {param['type']}"
+                    )
+            messages.append(
+                ChatMessageAssistant(
+                    content=content,
+                    tool_calls=tool_calls,
+                    model=model_name,
+                )
+            )
+
+            pending_assistant_message_params.clear()
+
+    for item in input:
+        # accumulate assistant message params until we clear the assistant message
+        if is_assistant_message_param(item):
+            pending_assistant_message_params.append(item)
+            continue
+
+        # see if we need to collect a pending assistant message
+        collect_pending_assistant_message()
+
+        if is_response_input_message(item):
+            # normalize item content
+            item_content: list[
+                ResponseInputTextParam
+                | ResponseInputImageParam
+                | ResponseInputFileParam
+            ] = (
+                [ResponseInputTextParam(type="input_text", text=item["content"])]
+                if isinstance(item["content"], str)
+                else item["content"]
+                if isinstance(item["content"], list)
+                else cast(
+                    list[
+                        ResponseInputTextParam
+                        | ResponseInputImageParam
+                        | ResponseInputFileParam
+                    ],
+                    [item["content"]],
+                )
+            )
+
+            # create inspect content
+            content = [
+                content_from_response_input_content_param(c) for c in item_content
+            ]
+            if item["role"] == "user":
+                messages.append(ChatMessageUser(content=content))
+            elif item["role"] == "assistant":
+                messages.append(ChatMessageAssistant(content=content))
+            else:
+                messages.append(ChatMessageSystem(content=content))
+        elif is_function_call_output(item):
+            messages.append(
+                ChatMessageTool(
+                    tool_call_id=item["call_id"],
+                    function=function_calls_by_id.get(item["call_id"]),
+                    content=[ContentText(text=item["output"])],
+                )
+            )
+        elif is_computer_call_output(item):
+            messages.append(
+                ChatMessageTool(
+                    tool_call_id=item["call_id"],
+                    function=function_calls_by_id.get(item["call_id"]),
+                    content=[ContentImage(image=item["output"]["image_url"])],
+                )
+            )
+        else:
+            # ImageGenerationCall
+            # ResponseCodeInterpreterToolCallParam
+            # McpApprovalRequest
+            # McpApprovalResponse
+            # ResponseCustomToolCallOutputParam
+            # ResponseCustomToolCallParam
+            # LocalShellCall
+            # LocalShellCallOutput
+            # ResponseFileSearchToolCallParam
+            # ItemReference
+            raise RuntimeError(
+                f"Type {item['type']} is not supported by the agent bridge"
+            )
+
+    # final collect of pending assistant message
+    collect_pending_assistant_message()
+
+    return messages
+
+
+output_item_adapter = TypeAdapter(list[ResponseOutputItem])
+
+
+def responses_output_items_from_assistant_message(
+    message: ChatMessageAssistant,
+) -> list[ResponseOutputItem]:
+    output: list[ResponseOutputItem] = []
+    for content in message.content:
+        if isinstance(content, ContentText):
+            # check for content.internal
+            if content.internal:
+                internal: str = f"\n{content_internal_tag(content.internal)}\n"
+            else:
+                internal = ""
+
+            # apply internal to content
+            content_text = f"{content.text}{internal}"
+
+            output.append(
+                ResponseOutputMessage(
+                    type="message",
+                    id=uuid(),
+                    role="assistant",
+                    content=[
+                        ResponseOutputRefusal(type="refusal", refusal=content_text)
+                        if content.refusal
+                        else ResponseOutputText(
+                            type="output_text", text=content_text, annotations=[]
+                        )
+                    ],
+                    status="completed",
+                )
+            )
+        elif isinstance(content, ContentReasoning):
+            reasoning = responses_reasoning_from_reasoning(content)
+            output.append(ResponseReasoningItem.model_validate(reasoning))
+
+        elif isinstance(content, ContentToolUse):
+            # we ignore these as they are encapsulated in AssistantInternal (which is necessary as there is no way to represent the full space of possible ContentToolUse in the responses API)
+            pass
+
+    for tool_call in message.tool_calls or []:
+        if tool_call.function == "computer":
+            output.append(
+                ResponseComputerToolCall(
+                    id=uuid(),
+                    type="computer_call",
+                    action=tool_call_arguments_to_action(tool_call.arguments),
+                    call_id=tool_call.id,
+                    pending_safety_checks=[],
+                    status="completed",
+                )
+            )
+        else:
+            output.append(
+                ResponseFunctionToolCall(
+                    type="function_call",
+                    call_id=tool_call.id,
+                    name=tool_call.function,
+                    arguments=json.dumps(tool_call.arguments),
+                )
+            )
+
+    return output
